Updated some functionality with Spectate
created a HashMap that stores user name and MATCH ID. This will be useful in telling whether we already know if someone is in game or not. Would say I'm about 60-70% done with the project. The project will be useable at 90% completion and the other 10% is quality of life and other fun things such as a way to track your bet record and a free currency to bet with.