HashMap Read + Write
When getSpectatorInfo() is ran it will check if the player that is passed through as a parameter is in the hash map using the .has function. The .has function returns true or false so if the player if it returns false then we add the player to hash map. Eventually the getSpectatorInfo() function will return 404 and if it does it checks to see if the player that is passed through the parameter is in the hash map. The only case that this is possible is if the player was in game and then got out of it. With this information we can access the hash map to check make an API request to the matchV4 (not implemented yet).